Added test label to Aarch64 machine
Currently the only Aarch64 machine is not able to compile tests, but the
testing label is being added to allow for the option of testing in the
future. The aarch64_linux and aarch64_linux_xl platforms are not included
in the target 'all', and therefore will not be run during the nightly
builds.

Signed-off-by: Colton Mills <millscolt3@gmail.com>