Change all nodes to mode NORMAL
Instead of 'Only build jobs with label
expressions matching this node' change
to 'Use this node as much as possible'.

This change was triggered because the test
jobs in post-parallel step now look for
'any' node. Which means they won't land on
any EXCLUSIVE node. The jobs have started
triggering the kubernetes cloud agents
to be provisioned.

Note: I left aarch64 as is because we only
have one machine for that platform.

Signed-off-by: Adam Brousseau <adam.brousseau88@gmail.com>