Create 'swt.natives' labels for eclipse.platform.releng Jenkins agents
As part of https://gitlab.eclipse.org/eclipsefdn/helpdesk/-/issues/2448
systematically named labels are added to those agents that are capable
to build the SWT native binaries for a certain platform. The label name
indicates for which platform the natives can be build by the agent and
therefore allow their use in a Jenkins matrix stage.

Signed-off-by: Hannes Wellmann <wellmann.hannes1@gmx.net>