feat: add github-scm-trait-notification-context plguin to JKube instance
In order to be able to customize the GitHub checks produced by the github-branch-source plugin, the github-scm-trait-notification-context needs to be present in order to be able to modify the other plugin's behavior.

Configuration of the check/notification message is especially critical when more than one job is run for a same PR, since the status/check of one Job/Pipeline will overwrite the status/check of the other one.

- https://plugins.jenkins.io/github-branch-source/
- https://plugins.jenkins.io/github-scm-trait-notification-context/
https://github.com/jenkinsci/github-scm-trait-notification-context-plugin