Link native build nodes into eclipse.platform and equinox JIPP
Besides linking the 'jenkins/configuration.yml' to the eclipse.platform
and equinox JIPP, this introduce the 'remoteFSSuffix' configuration,
which is replaced with an project specific value while building images.
This is used to ensure different remoteFS locations for
the different JIPPs that use the native nodes.

The value of project.fullName is not used to keep paths shorter and to
avoid issues on Windows regarding too long paths.

Required for
https://gitlab.eclipse.org/eclipsefdn/helpdesk/-/issues/4231